+++ azplugins/PairPotentials.cu
@@ -28,5 +28,10 @@ template cudaError_t compute_pair_potential<azplugins::detail::PairEvaluatorColl
     (const pair_args_t& pair_args,
      const typename azplugins::detail::PairEvaluatorColloid::param_type *d_params);
 
+//! Kernel driver for core-shifted Lennard-Jones pair potential
+template cudaError_t compute_pair_potential<azplugins::detail::PairEvaluatorShiftedLJ>
+    (const pair_args_t& pair_args,
+     const typename azplugins::detail::PairEvaluatorShiftedLJ::param_type *d_params);
+
 } // end namespace gpu
 } // end namespace azplugins